requirements.txt
flask
flask_restful
flask_bootstrap
Pillow
numpy
requests